+++ lib/gg_local_package_dependencies.dart
@@ -9,4 +9,5 @@ library;
 export 'src/gg_local_package_dependencies.dart';
 export 'src/commands/graph.dart';
 export 'src/commands/processing_list.dart';
+export 'src/commands/sorted_processing_list.dart';
 export 'src/tools/node.dart';